+++ CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.3.0] - 2026-04-24
+
+### Added
+
+- Library: "parallel" feature triggers rayon (opt-in).
+- CLI: sets parallel feature.
+
+### Changed
+
+- Intermediates logic based on Flatbush and Concave Hull algorithms.
+
 ## [0.2.0] - 2026-04-22
 
 ### Added